qol improvements

+++ src/app/modules/coach/pages/coach-courses-list/coach-courses-list.component.ts
@@ -2,18 +2,24 @@ import { Component, OnInit } from '@angular/core';
 import { RouterLink } from '@angular/router';
 import { Course } from '../../models/Course';
 import { CourseService } from '../../service/course.service';
+import { ConfirmDialogModule } from 'primeng/confirmdialog';
+import { ConfirmationService } from 'primeng/api';
 
 @Component({
   selector: 'app-coach-courses-list',
   standalone: true,
-  imports: [RouterLink],
+  imports: [RouterLink, ConfirmDialogModule],
+  providers: [ConfirmationService],
   templateUrl: './coach-courses-list.component.html',
   styleUrl: './coach-courses-list.component.scss'
 })
 export class CoachCoursesListComponent implements OnInit {
   protected courses: Course[] = [];
 
-  public constructor(private courseService: CourseService) {}
+  public constructor(
+    private courseService: CourseService,
+    private confirmationService: ConfirmationService
+  ) {}
 
   public ngOnInit(): void {
     this.getAll();
@@ -26,8 +32,15 @@ export class CoachCoursesListComponent implements OnInit {
   }
 
   protected deleteCourse(id: string): void {
-    this.courseService.delete(id).subscribe(() => {
-      this.getAll();
+    this.confirmationService.confirm({
+      message: 'Are you sure you want to delete this course?',
+      header: 'Confirm',
+      accept: () => {
+        this.courseService.delete(id).subscribe(() => {
+          this.getAll();
+        });
+      },
+      reject: () => {}
     });
   }
 }
